Fixed mutation race condition
