+++ app/client/src/ce/actions/evaluationActions.ts
@@ -66,8 +66,6 @@ export const EVALUATE_REDUX_ACTIONS = [
   ReduxActionTypes.MOVE_ACTION_SUCCESS,
   ReduxActionTypes.RUN_ACTION_SUCCESS,
   ReduxActionErrorTypes.RUN_ACTION_ERROR,
-  ReduxActionTypes.EXECUTE_PLUGIN_ACTION_SUCCESS,
-  ReduxActionErrorTypes.EXECUTE_PLUGIN_ACTION_ERROR,
   ReduxActionTypes.CLEAR_ACTION_RESPONSE,
   // JS Actions
   ReduxActionTypes.CREATE_JS_ACTION_SUCCESS,
