+++ app/client/src/workers/Evaluation/getEntityForContext.ts
@@ -2,7 +2,6 @@ import type { JSActionEntity } from "@appsmith/entities/DataTree/types";
 import type { DataTreeEntity } from "entities/DataTree/dataTreeTypes";
 import { ENTITY_TYPE_VALUE } from "entities/DataTree/dataTreeFactory";
 import JSObjectCollection from "./JSObject/Collection";
-import JSFactory from "./JSObject/JSVariableFactory";
 import { jsObjectFunctionFactory } from "./fns/utils/jsObjectFnFactory";
 import { isObject } from "lodash";
 
@@ -53,7 +52,8 @@ export function getEntityForEvalContext(
           return Object.assign({}, jsObject, fns);
         }
 
-        jsObjectForEval = JSFactory.create(entityName, jsObjectForEval);
+        jsObjectForEval =
+          JSObjectCollection.getVariablesForEvaluationContext(entityName);
         return Object.assign(jsObjectForEval, fns);
       }
     }
